Sounds (#53)
* warning fixes

* adapting to newer soundengine

* experimental Asteroids sound effects

* adaptions to new soundengine

* evolving sound implementation

* systems factory and some trials

TODO: fix quit from menu

* scene and sound management

* minimal restucture and loading screen

* loading scene with progress bar

* working pre-load

* Some tidy up of LambdaSystem

* warning fixes

* some ideas for progress bar

* alternative progress

* build and warning fixes

* foundations library

* moving foundation namespaces

* warning fixes

* more foundation

* using resourceLocator

---------

+++ tests/Vec/VecMathTests.cpp
@@ -13,7 +13,7 @@ TEST_CASE("angleBetween", "[VecTests]")
     auto  data = GENERATE(table<pg::fVec2, float>({
         {{0, 1}, 0.0f},
         {{1, 0}, -std::numbers::pi / 2.0f},
-        {{0, -1}, -std::numbers::pi},
+        {{0, -1}, static_cast<float>(-std::numbers::pi)},
         {{-1, 0}, std::numbers::pi / 2.0f},
     }));
